% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{auto_reformat}
\alias{auto_reformat}
\title{auto_reformat}
\usage{
auto_reformat(
  df,
  prefix = "",
  rename.cols = TRUE,
  reorder.rows = TRUE,
  reorder.cols = TRUE,
  strip.html = FALSE
)
}
\arguments{
\item{df}{data.frame to rename and reorder columns}

\item{prefix}{Optional prefix to add before renamed columns}
}
\value{
DF with renamed and reordered columns
}
\description{
Automatically rename, reorder, according to map and optionally
strip HTML from relevant columns
}
