% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_duplicate_question}
\alias{check_duplicate_question}
\title{check_duplicate_question}
\usage{
check_duplicate_question(survey)
}
\arguments{
\item{survey}{Survey to check}
}
\description{
Check whether a survey has a duplicate question number.
}
