% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_duplicate_questions}
\alias{check_duplicate_questions}
\title{check_duplicate_questions}
\usage{
check_duplicate_questions(survey, fatal = FALSE)
}
\arguments{
\item{survey}{Survey object to check}

\item{fatal}{Stop execution if duplicate question, or no}
}
\description{
Check whether a survey has a duplicate question number and report
to user if it does
}
