% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{choice_to_factor}
\alias{choice_to_factor}
\title{choice_to_factor}
\usage{
choice_to_factor(df, choice_rev = FALSE)
}
\arguments{
\item{choice_rev}{BOOL, whether you want to SWITCH the order of factored choices. REQUIRES factor = TRUE to be set DEFAULT = FALSE}

\item{choices_df}{choices sub object (DF) pulled from the qsurvey object.}

\item{choice_factor}{BOOL, whether you want choicse to be a factor or not DEFAULT = FALSE}
}
\value{
data.frame of choices ready to be merged with the main question df
}
\description{
Cleaned data.frame of question choicses ready for plotting. This function
works for regular matrix and multiple choices questions. It calls a SBS matrix specific
function if a sbs matrix question is provided
}
