% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_funs.R
\name{clean_resp_sbs}
\alias{clean_resp_sbs}
\title{FUN: clean_resp_sbs}
\usage{
clean_resp_sbs(df)
}
\arguments{
\item{df}{a df object derived from Qtoolkit with the question and response values}
}
\value{
data.frame of choices ready to be merged
}
\description{
This function cleans up the columns in a SBS matrix so that is it really for 
joining and final plotting
}
