% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{connect_qualtrics}
\alias{connect_qualtrics}
\title{connect_qualtrics}
\usage{
connect_qualtrics(subdomain, file = "~/.qualtrics_api_key", key = FALSE)
}
\arguments{
\item{subdomain}{Qualtrics subdomain with which to get surveys}

\item{file}{Filename of file with connection information}

\item{key}{Qualtrics API key for API access}
}
\description{
Open a connection to Qualtrics API with login info
}
