% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{empty_to_na}
\alias{empty_to_na}
\title{empty_to_na}
\usage{
empty_to_na(lst)
}
\arguments{
\item{list}{List to test}
}
\value{
original list with empty elements NA
}
\description{
Loop thru list and turn empty dataframes and lists into NA
}
