% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{gather_responses}
\alias{gather_responses}
\title{gather_responses}
\usage{
gather_responses(quest_obj)
}
\arguments{
\item{quest_obj}{the object within the qsurvey object that contains all relevant information to the question including choicses and subquestions IF those are relevant}
}
\value{
Nice data.frame ready for pretty plotting.
}
\description{
Cleaned data.frame all ready for plotting
}
