% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{get_choices_sbs}
\alias{get_choices_sbs}
\title{FUN: get_choices_sbs}
\usage{
get_choices_sbs(choices_df, choice_wrap = NULL, choice_factor = FALSE,
  choice_rev = FALSE)
}
\arguments{
\item{choices_df}{choices sub object (DF) pulled from the qsurvey object.}

\item{choice_wrap}{int - the number of characters at which you want to wrap a line, DEFAULT = NULL}

\item{choice_factor}{BOOL, whether you want choicse to be a factor or not DEFAULT = FALSE}

\item{choice_rev}{BOOL, whether you want to SWITCH the order of factored choices. REQUIRES factor = TRUE to be set DEFAULT = FALSE}
}
\value{
data.frame of choices ready to be merged with the main question df with ordered factors
applied if selected
}
\description{
This function is specific for side by side matrices. These survey elemtns are more complex
than regular matrices because they contain several sub matrices and thus a different
structure that needs to be accounted for. This function handles factors associated with 
choices in each matrix that may or may not be the same.
}
