% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_question}
\alias{get_question}
\title{get_question}
\usage{
get_question(survey, question_name = "Q", question_id = "QID")
}
\arguments{
\item{survey}{Survey Design Object}

\item{question_name}{Question name (ie "Q10")}

\item{question_id}{Question ID (ie "QID9")}
}
\value{
DF of matching questions
}
\description{
Get info about question(s) asked in a particular survey
}
