% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{get_question_resp}
\alias{get_question_resp}
\title{get_data}
\usage{
get_question_resp(quest_obj)
}
\arguments{
\item{quest_obj}{the object within the qsurvey object that contains all relevant information to the question including choicses and subquestions IF those are relevant}
}
\value{
Nice data.frame ready for plotting. yeaaassss
}
\description{
Cleaned data.frame all ready for plotting
}
