% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_questions}
\alias{get_questions}
\title{get_questions}
\usage{
get_questions(survey, question = "Q")
}
\arguments{
\item{survey}{Survey Design Object}

\item{question}{Question Number}
}
\value{
DF of matching questions
}
\description{
Get info about questions asked in a particular survey
}
