% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_responses}
\alias{get_responses}
\title{get_responses}
\usage{
get_responses(survey, question = "Q", metadata = FALSE)
}
\arguments{
\item{survey}{Survey Object (or ID)}

\item{question}{Question Number}

\item{metadata}{Include metadata about each participant in DF}
}
\value{
DF of survey question answers
}
\description{
Get responses for a particular survey question
}
