% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_funs.R
\name{get_SBS}
\alias{get_SBS}
\title{FUN: get_sbs}
\usage{
get_SBS(
  quest_obj,
  quest_wrap = NULL,
  choice_wrap = NULL,
  choice_factor = FALSE,
  choice_rev = FALSE
)
}
\arguments{
\item{quest_obj}{the object within the qsurvey object that contains all relevant information to the question including choicses and subquestions IF those are relevant}

\item{quest_wrap}{an integer value defining the number of characters at which you want to specify a line break for each question at for prettier plotting. default = NULL which means no wrap}

\item{choice_wrap}{an integer value defining the number of characters at which you want to specify a line break for each CHOICE at for prettier plotting. default = NULL which means no wrap}

\item{choice_factor}{boolean set to TRUE if you want choices to be turned into a factor and ordered by the order provided in qualtrics (this is the order that you provided to the user in the survey)}

\item{choice_rev}{boolean set to TRUE if you want to reverse the order of choices}
}
\value{
data.frame of choices ready to be merged
}
\description{
This function cleans up the columns in a SBS matrix so that is it really for 
joining and final plotting. If the matrices have difference choice options, AND factors are set to TRUE
then it returns a list containing a DF for each sub matrix. Otherwise it returns a regular df. 
It's not ideal for a function to return 2 diff data types but in this instance, there may not be a better way if you want factors
}
