% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_survey}
\alias{get_survey}
\title{get_survey}
\usage{
get_survey(filter = "", match.exact = TRUE, verbose = TRUE)
}
\arguments{
\item{filter}{DF of surveys with survey ids and names, or survey name}

\item{match.exact}{Match exact survey name or partial}

\item{verbose}{User will see which surveys are being loaded}
}
\value{
If one match, return design object of survey. If many, return list of design objects
}
\description{
Get all metadata of specific survey(s) from API and return
survey design object
}
