% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_survey_id_by_name}
\alias{get_survey_id_by_name}
\title{get_survey_id_by_name}
\usage{
get_survey_id_by_name(name, match.exact = TRUE, surveys_df = NULL)
}
\arguments{
\item{name}{Survey name}

\item{match.exact}{Search for exact name? (or partial)}

\item{surveys_df}{Dataframe of available surveys. If NULL, will load search surveys loaded from Qualtrics API}
}
\value{
Survey ID
}
\description{
Get a survey ID by name
}
