% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_surveys}
\alias{get_surveys}
\title{get_surveys}
\usage{
get_surveys(filter = "", match.exact = FALSE)
}
\arguments{
\item{filter}{Complete or partial name of survey}

\item{match.exact}{Match exact survey name or partial}
}
\value{
DF of matched surveys
}
\description{
Select surveys available, optionally matching survey name or ID.
}
