% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.obj.type}
\alias{is.obj.type}
\title{is.obj.type}
\usage{
is.obj.type(var, type)
}
\arguments{
\item{var}{Object to be tested}

\item{type}{Type of object to be tested}
}
\value{
TRUE if object is qsurvey object; otherwise FALSE
}
\description{
Test if passed variable is particular object
}
