% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nested_list_to_df}
\alias{nested_list_to_df}
\title{nested_list_to_df}
\usage{
nested_list_to_df(lst)
}
\arguments{
\item{lst}{Named list}
}
\value{
Dataframe instead of nested List
}
\description{
Convert a named list to dataframe and preserve type
}
