% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_connect}
\alias{qapi_connect}
\title{qapi_connect}
\usage{
qapi_connect(org_id, api_key, auth_file = ".qapi_auth.R", verbose = FALSE)
}
\arguments{
\item{org_id}{Qualtrics org_id with which to get surveys}

\item{api_key}{Qualtrics API key}

\item{auth_file}{File from which to source Qualtrics API auth info}
}
\description{
Open a connection to Qualtrics API with login info
}
