% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_error}
\alias{qapi_error}
\title{qapi_error}
\usage{
qapi_error(request)
}
\arguments{
\item{request}{httr request object of the Qualtrics API request}
}
\description{
Handle errors caused by Qualtrics API request
}
