% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_get_auth}
\alias{qapi_get_auth}
\title{qapi_get_auth}
\usage{
qapi_get_auth()
}
\value{
Named list of authentication parameters
}
\description{
Get the stored authentication parameters for Qualtrics API
}
