% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_get_base_url}
\alias{qapi_get_base_url}
\title{qapi_base_url}
\usage{
qapi_get_base_url(subdomain)
}
\arguments{
\item{subdomain}{Qualtrics subdomain}
}
\value{
String of Qualtrics API base URL
}
\description{
Return a Qualtrics API base URL based upon subdomain
}
