% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_get_survey}
\alias{qapi_get_survey}
\title{qapi_get_survey}
\usage{
qapi_get_survey(survey_id)
}
\arguments{
\item{survey_id}{}
}
\value{
Named list of metadata
}
\description{
QAPI call to get metadata about a particular survey
https://api.qualtrics.com/docs/get-survey
}
