% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_request}
\alias{qapi_request}
\title{qapi_request}
\usage{
qapi_request(verb, method, data = list(), content.as = "text",
  auth = NULL, all.results = TRUE)
}
\arguments{
\item{verb}{Request type (GET, POST, ...)}

\item{method}{API call method (surveys, reponseexports, ...) or full API URL}

\item{data}{Named list with request payload data}

\item{content.as}{"text" or "raw" depending on if JSON or raw data returned}

\item{auth}{Qualtrics API authentication to use}

\item{all.results}{Return all results if paginated, or just one page}
}
\value{
Named list of JSON decoded response content
}
\description{
Send request to Qualtrics API
}
