% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_response_export}
\alias{qapi_response_export}
\title{qapi_response_export}
\usage{
qapi_response_export(survey_id)
}
\arguments{
\item{survey_id}{ID of survey to get responses}
}
\value{
DF of survey responses
}
\description{
Get DF of survey responses from Qualtrics API
}
