% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_test}
\alias{qapi_test}
\title{qapi_test}
\usage{
qapi_test(org_id, key, verbose = FALSE)
}
\arguments{
\item{org_id}{Qualtrics org_id to test}

\item{key}{Qualtrics API key to test}
}
\value{
True if connection successful; error if not
}
\description{
Test Qualtrics API connection
}
