% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qquestion.R
\name{qquestion}
\alias{qquestion}
\title{qquestion}
\usage{
qquestion(q_meta, q_resp, clean_html)
}
\arguments{
\item{q_meta}{Named list of question metadata}

\item{q_resp}{DF of question responses}

\item{clean_html}{BOOLEAN - determines whether to strip the html surrounding text or not}
}
\value{
Qualtrics Survey Question object
}
\description{
Qualtrics Survey Question object
}
