% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsurvey.R
\name{qsurvey}
\alias{qsurvey}
\title{qsurvey}
\usage{
qsurvey(id_or_name, strip_html = TRUE, include.raw = FALSE)
}
\arguments{
\item{id_or_name}{Survey ID or name}

\item{include.raw}{Include raw JSON in survey object}

\item{strip.html}{Auto-strip HTML from question names & choices}
}
\value{
Qualtrics survey object
}
\description{
Qualtrics survey object
}
