% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{split_strings}
\alias{split_strings}
\title{Wrap long strings at x characters for pretty plotting and data viz.}
\usage{
split_strings(string, nchar = 35)
}
\arguments{
\item{string}{of type char: the string that needs to be wrapped}

\item{n_char}{number of characters to insert a \n after - default: 35}
}
\value{
a string with line breaks at every n_char characters
}
\description{
Wrap long strings at x characters for pretty plotting and data viz.
}
\examples{
a_long_string <- "this is a string that i'd like to wrap"
split_strings(a_long_string, n_char = 15)
}
