% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{strip_html}
\alias{strip_html}
\title{strip_html}
\usage{
strip_html(text, consolidate = TRUE)
}
\arguments{
\item{text}{Text to strip HTML from}

\item{consolidate}{Consolidate stripped HTML; Turn line break into a space, collapse extra spaces.}
}
\value{
Text with HTML stripped
}
\description{
Strip HTML tags from string, with RegEx
}
