% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{type_acronym_to_text}
\alias{type_acronym_to_text}
\title{type_acronym_to_text}
\usage{
type_acronym_to_text(acronym)
}
\arguments{
\item{acronym}{Acronym to convert to human readable text}
}
\value{
Long text form of acronym if acronym is known, or empty string
}
\description{
Convert acronym of question type/selector/subselector to human readable
}
