% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_choices}
\alias{get_choices}
\title{get_choices}
\usage{
get_choices(survey, question_num = "Q.*", question_id = "QID.*")
}
\arguments{
\item{survey}{Survey Design Object}

\item{question_num}{Question number (ie "Q10")}

\item{question_id}{Question ID (ie "QID9")}
}
\value{
DF of choices for each question
}
\description{
Get possible choices for a particular question
}
