% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{get_choices}
\alias{get_choices}
\title{get_choices}
\usage{
get_choices(choices_df, choice_wrap = NULL, choice_factor = FALSE,
  choice_rev = FALSE)
}
\arguments{
\item{choices_df}{choices sub object pulled from the qsurvey object.}
}
\value{
data.frame of choices ready to be merged with the main question df
}
\description{
Cleaned data.frame all ready for plotting
}
