% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_funs.R
\name{get_choices_sbs}
\alias{get_choices_sbs}
\title{FUN: get_choices_sbs}
\usage{
get_choices_sbs(choices_df, choice_wrap = NULL, choice_factor = FALSE,
  choice_rev = FALSE)
}
\arguments{
\item{choices_df}{choices sub object (DF) pulled from the qsurvey object.}

\item{choice_wrap}{int - the number of characters at which you want to wrap a line, DEFAULT = NULL}

\item{choice_factor}{BOOL, whether you want choicse to be a factor or not DEFAULT = FALSE}

\item{choice_rev}{BOOL, whether you want to SWITCH the order of factored choices. REQUIRES factor = TRUE to be set DEFAULT = FALSE}
}
\value{
data.frame of choices ready to be merged with the main question df with ordered factors
applied if selected
}
\description{
This function is specific for side by side (SBS) matrices. SBS matrices have a col
column which dilineates the sub matrix that the question choicses are in
Some use cases will have the same choicses in each submatrix. Others will have different
choices making the work flow more complex
}
