% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_questions}
\alias{get_questions}
\title{get_questions}
\usage{
get_questions(survey, question_num = ".*", question_id = "QID.*")
}
\arguments{
\item{survey}{Survey Design Object}

\item{question_num}{Question number (ie "Q10")}

\item{question_id}{Question ID (ie "QID9")}
}
\value{
DF of matching questions
}
\description{
Get info about question(s) asked in a particular survey
}
