% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{get_subq}
\alias{get_subq}
\title{get_subq}
\usage{
get_subq(subq_df)
}
\arguments{
\item{subq_df}{choicse sub object pulled from the qsurvey object.}
}
\value{
data.frame of subquestions ready to be merged with the main question df
}
\description{
Cleaned data.frame all ready for plotting
}
