% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_survey}
\alias{get_survey}
\title{get_survey}
\usage{
get_survey(name = "", id = "")
}
\arguments{
\item{name}{Complete or partial name of survey}
}
\value{
DF of matched surveys
}
\description{
Select surveys available, optionally filtering by name
}
