% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{load_survey}
\alias{load_survey}
\title{load_survey}
\usage{
load_survey(surveys = "", verbose = TRUE, match.exact = TRUE)
}
\arguments{
\item{surveys}{DF of surveys with survey ids and names, or survey name}

\item{verbose}{Let user know full name & surveyID of selected survey}

\item{match.exact}{Match exact survey name or use wildcards}
}
\value{
If one match, return design object of survey. If many, return list of design objects
}
\description{
Select surveys with a fragment of the survey's name
}
