% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_connect}
\alias{qapi_connect}
\title{qapi_connect}
\usage{
qapi_connect(subdomain, key, auth_file = ".qapi_auth.R")
}
\arguments{
\item{subdomain}{Qualtrics subdomain with which to get surveys}

\item{key}{Qualtrics API key for API access}

\item{file}{File from which to load Qualtrics API Auth info}
}
\description{
Open a connection to Qualtrics API with login info
}
