% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qapi_test}
\alias{qapi_test}
\title{qapi_test}
\usage{
qapi_test(subdomain, key)
}
\arguments{
\item{subdomain}{Qualtrics subdomain to test}

\item{key}{Qualtrics API key to test}
}
\value{
True if successful, or will error if not
}
\description{
Test Qualtrics API connection
}
