% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{split_strings}
\alias{split_strings}
\title{split_strings}
\usage{
split_strings(string, nchar = 35)
}
\arguments{
\item{string}{of type char: the string that needs to be wrapped}

\item{nchar}{number of characters to insert a line break after - default: 35}
}
\value{
a string with line breaks at every n_char characters
}
\description{
Wrap long strings at x characters for pretty plotting and data viz.
}
